Singapore-23/requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
PyJWT==2.8.0
pyngrok==7.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
twilio==8.10.1
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
